simple-microservice/requirements.txt
# Závislosti - Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
prometheus-client==0.19.0